// src/frontend/api/responseTypes.ts
export interface RainwaveResponseTypes {
  album: AlbumWithDetail;
  album_diff: AlbumDiff;
  albums: SearchResult['albums'];
  all_albums_paginated: AllAlbumsPaginated;
  all_artists_paginated: AllArtistsPaginated;
  all_faves: FaveSong[];
  all_groups_paginated: AllGroupsPaginated;
  all_songs: AllSongsSong[];
  all_stations_info: AllStationsInfo;
  already_voted: AlreadyVoted;
  api_info: ApiInfo;
  artists: SearchResult['artists'];
  artist: ArtistWithSongs;
  build_version: number;
  cookie_domain: string;
  delete_request_result: BooleanResult;
  error_report_result: BooleanResult;
  error: RainwaveErrorObject;
  fave_album_result: FaveAlbumResult;
  fave_all_songs_result: FaveAllSongsResult;
  fave_song_result: FaveSongResult;
  group: GroupWithDetail;
  listener: Listener;
  live_voting: LiveVoting;
  locale: string;
  locales: Record<string, string>;
  message_id: MessageId;
  mobile: boolean;
  on_draw: unknown[];
  on_init: unknown[];
  on_measure: unknown[];
  order_requests_result: BooleanResult;
  pause_request_queue_result: BooleanResult;
  ping: Ping;
  pong: Pong;
  pongConfirm: PongConfirm;
  playback_history: PlaybackHistory;
  rate_result: RateResult;
  redownload_m3u: RedownloadM3u;
  relays: Relays;
  request_favorited_songs_result: BooleanResult;
  request_line: RequestLine;
  request_result: BooleanResult;
  request_unrated_songs_result: BooleanResult;
  requests: Requests;
  sched_current: RainwaveEvent;
  sched_history: RainwaveEvent[];
  sched_next: RainwaveEvent[];
  sdk_exception: Error;
  sdk_error_clear: RainwaveSDKErrorClear;
  sdk_schedule_synced: true;
  song: SongWithDetail;
  songs: SearchResult['songs'];
  station_list: Record<Station, { id: number; url: string }>;
  station_song_count: StationSongCount;
  stations: Stations;
  stream_filename: string;
  sync_result: RainwaveErrorObject;
  top_100: Top100;
  unpause_request_queue_result: BooleanResult;
  unrated_songs: UnratedSongs;
  user_info: User;
  user_recent_votes: UserRecentVotes;
  user_requested_history: UserRecentVotes;
  user: User;
  vote_result: VoteResult;
  websocket_host: string;
  wsok: true;
  wsthrottle: { tl_key: string; text: string };
  wserror: RainwaveErrorObject;
}